+++ WHATSNEW.md
@@ -1,9 +1,11 @@
 ### Added
 
-- Option `--smtp` for `fmg assign`
+- Option `--smtp` for `fmg assign` and `fmg_post`
 - Method `inventory.get_item()` to only get one single item from the inventory
 
 ### Changed
+
+- `fortimanager.post()` now returns list of errors instead of just the number of errors
 
 ### Removed