+++ WHATSNEW.md
@@ -3,6 +3,11 @@
 - Option `--smtp` for `fmg assign`
 - Method `inventory.get_item()` to only get one single item from the inventory
 
+### Fixed
+
+- Result.print_result_as_table() gives error with empty Response. Now it just prints an empty line
+if no results were pushed.
+
 ### Changed
 
 ### Removed