+++ WHATSNEW.md
@@ -1,6 +1,6 @@
 ### Added
 
-- Option `--smtp` for `fmg assign`
+- Option `--smtp` for `fmg assign` and `fmg_post`
 - Method `inventory.get_item()` to only get one single item from the inventory
 
 ### Fixed
@@ -9,6 +9,8 @@
 if no results were pushed.
 
 ### Changed
+
+- `fortimanager.post()` now returns list of errors instead of just the number of errors
 
 ### Removed