+++ WHATSNEW.md
@@ -5,6 +5,7 @@
 
 ### Fixed
 
+- Update to urllib3 2.0.6 due to dependabot security warning
 - No more Traceback for FortiGates without hostname or token specified
 - Result.print_result_as_table() gives error with empty Response. Now it just prints an empty line
 if no results were pushed.